Import a dataset

This import function can import data from delimited text files, EXCEL spreadsheet, and statistical packages such as SAS, SPSS, and Stata.

import(file, ...)

Arguments

file

data file to import.

...

parameters passed to the import function.

Value

a data frame

Details

The import function is a wrapper for the haven, readxl, rjson, and vroom packages.

Note

Complex nested JSON files will not be imported properly.

See also

Examples

if (FALSE) {

#import a comma delimited file
mydataframe <- import("mydata.csv")

#import a SAS binary file
mydataframe <- import("mydata.sas7bdat")

#import the second worksheet of an Excel workbook
mydataframe <- import("mydata.xlsx", sheet = 2)

#prompt for a file to import
mydataframe <- import()
}